requirements.txt
Chameleon==2.8.4
Mako==0.7.0
MarkupSafe==0.15
PasteDeploy==1.5.0
Pygments==1.5
SQLAlchemy==0.7.6
WebOb==1.2b3
distribute==0.6.19
lazr.restfulclient==0.11.2
lazr.uri==1.0.2
oauth==1.0.1
pexpect==2.3
psycopg2==2.4.2
pyOpenSSL==0.12
pycrypto==2.3
pycurl==7.19.0
pyramid==1.3
pyramid-debugtoolbar==1.0.2
pyramid-tm==0.4
pyserial==2.5
repoze.lru==0.5
simplejson==2.1.6
smart==1.4
transaction==1.2.0
translationstring==1.1
venusian==1.0a6
virtualenv==1.6.4
virtualenvwrapper==3.2
wadllib==1.2.0
waitress==0.8.1
wsgiref==0.1.2
zope.deprecation==3.5.1
zope.interface==3.8.0
zope.sqlalchemy==0.7
